app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
Pillow==10.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
greenlet>=3.0.0
